Progress before rewrite
I've run into several issues with BS and have realised the bootloader
needs essentially a full rewrite. I need more than 1mb of memory to
load the kernel, which means the bootloader itself needs to be in x64
mode. This means the boot sequence is actually going to need 3 stages -
the bootstrapper, code to enter 64-bit mode, and then the bootloader.
Doing this is going to require a lot of code changes.